helpdesk/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
Pillow>=9.5.0
gunicorn>=21.2.0
whitenoise>=6.6.0
